refactor(backend): unify duplicate AdminGuard behind one canonical implementation
