+++ xconfess-backend/src/admin/admin.controller.spec.ts
@@ -3,7 +3,7 @@ import { AdminController } from './admin.controller';
 import { AdminService } from './services/admin.service';
 import { ModerationService } from './services/moderation.service';
 import { JwtAuthGuard } from '../auth/jwt-auth.guard';
-import { AdminGuard } from './guards/admin.guard';
+import { AdminGuard } from '../auth/admin.guard';
 
 describe('AdminController', () => {
   let controller: AdminController;
