+++ xconfess-backend/src/admin/guards/admin.guard.spec.ts
@@ -1,9 +1,10 @@
-import { AdminGuard } from './admin.guard';
+import { AdminGuard } from '../../auth/admin.guard';
 import { ForbiddenException } from '@nestjs/common';
+import { UserRole } from '../../user/entities/user.entity';
 
 describe('AdminGuard', () => {
   it('throws if no user', () => {
-    const guard = new AdminGuard({} as any);
+    const guard = new AdminGuard();
     const ctx: any = {
       switchToHttp: () => ({
         getRequest: () => ({ user: null }),
@@ -13,20 +14,20 @@ describe('AdminGuard', () => {
   });
 
   it('throws if not admin', () => {
-    const guard = new AdminGuard({} as any);
+    const guard = new AdminGuard();
     const ctx: any = {
       switchToHttp: () => ({
-        getRequest: () => ({ user: { isAdmin: false } }),
+        getRequest: () => ({ user: { role: UserRole.USER } }),
       }),
     };
     expect(() => guard.canActivate(ctx)).toThrow(ForbiddenException);
   });
 
   it('allows admin', () => {
-    const guard = new AdminGuard({} as any);
+    const guard = new AdminGuard();
     const ctx: any = {
       switchToHttp: () => ({
-        getRequest: () => ({ user: { isAdmin: true } }),
+        getRequest: () => ({ user: { role: UserRole.ADMIN } }),
       }),
     };
     expect(guard.canActivate(ctx)).toBe(true);
