--- xconfess-backend/src/admin/guards/admin.guard.ts
@@ -1,26 +1,0 @@
-import {
-  Injectable,
-  CanActivate,
-  ExecutionContext,
-  ForbiddenException,
-} from '@nestjs/common';
-import { UserRole } from 'src/user/entities/user.entity';
-import { AuthenticatedRequest } from '../../auth/interfaces/jwt-payload.interface';
-
-@Injectable()
-export class AdminGuard implements CanActivate {
-  canActivate(context: ExecutionContext): boolean {
-    const request = context.switchToHttp().getRequest<AuthenticatedRequest>();
-    const user = request.user;
-
-    if (!user) {
-      throw new ForbiddenException('Authentication required');
-    }
-
-    if (user.role !== UserRole.ADMIN) {
-      throw new ForbiddenException('Admin access required');
-    }
-
-    return true;
-  }
-}
