+++ xconfess-backend/src/auth/admin.guard.ts
@@ -5,20 +5,17 @@ import {
   ForbiddenException,
 } from '@nestjs/common';
 import { UserRole } from '../user/entities/user.entity';
-import { Request } from 'express';
+import { AuthenticatedRequest } from './interfaces/jwt-payload.interface';
 
-type AuthenticatedRequest = Request & {
-  user?: {
-    role?: UserRole;
-  };
-};
-
+/**
+ * AdminGuard — single source of truth for admin authorization across all HTTP and WebSocket routes.
+ * Rejects unauthenticated requests and any authenticated user whose role is not ADMIN.
+ */
 @Injectable()
 export class AdminGuard implements CanActivate {
   canActivate(context: ExecutionContext): boolean {
     const request = context.switchToHttp().getRequest<AuthenticatedRequest>();
 
-    // Check if user is authenticated
     if (!request.user) {
       throw new ForbiddenException('User is not authenticated');
     }
@@ -26,7 +23,6 @@ export class AdminGuard implements CanActivate {
     const userRole = String(request.user.role || '').toLowerCase();
     const isAdmin = userRole === UserRole.ADMIN || userRole === 'admin';
 
-    // Check if user has admin role
     if (!isAdmin) {
       throw new ForbiddenException('Only admins can access this endpoint');
     }
